requirements.txt
# --- Core IA & Vision (Versiones fijas para evitar conflictos con RTDetrImageProcessor) ---
torch>=2.4.1
torchvision>=0.19.1
transformers>=4.48.0
timm>=1.0.11
accelerate>=1.1.1
docling>=2.83.0
pypdf==4.0.1  # Detección inteligente de PDFs scaneados vs texto
PyMuPDF>=1.24.0 # (fitz) Extracción multimodal de imágenes para Gemini

# --- Web Framework & API ---
fastapi==0.115.12
uvicorn==0.34.0
pydantic==2.11.1
pydantic-settings==2.8.1
python-multipart==0.0.20
aiofiles==24.1.0
python-dotenv==1.1.0

# --- Async & Background Tasks ---
celery==5.5.0
redis==5.2.1

# --- Database & Storage ---
asyncpg==0.30.0
psycopg2-binary==2.9.10
sqlalchemy==2.0.40
minio==7.2.15

# --- LLM Integrations ---
langchain==0.3.21
langchain-openai==0.3.5
langchain-google-genai==2.0.1
langchain-ollama==0.2.3
